Clean up packeging
