+++ app/src/main/java/com/gathering/android/event/myevent/view/MyEventAdapter.kt
@@ -1,4 +1,4 @@
-package com.gathering.android.event.home
+package com.gathering.android.event.myevent.view
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -6,55 +6,54 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.gathering.android.common.ImageLoader
-import com.gathering.android.databinding.ItemEventBinding
+import com.gathering.android.databinding.ItemMyEventBinding
 import com.gathering.android.event.model.Event
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.android.scopes.ActivityScoped
 import javax.inject.Inject
 
 @ActivityScoped
-class EventListAdapter @Inject constructor(
+class MyEventAdapter @Inject constructor(
     @ApplicationContext private val context: Context,
     private val imageLoader: ImageLoader
-) : RecyclerView.Adapter<EventListAdapter.ViewHolder>() {
+) : RecyclerView.Adapter<MyEventAdapter.ViewHolder>() {
 
-    private var eventItemList: MutableList<Event> = mutableListOf()
-    private var onEventClicked: (event: Event) -> Unit = {}
+    private var myEventItemList: MutableList<Event> = mutableListOf()
+    private var onMyEventClicked: (event: Event) -> Unit = {}
     private val li = LayoutInflater.from(context)
 
+
     @SuppressLint("NotifyDataSetChanged")
-    fun setEventItem(eventItemList: MutableList<Event>) {
-        this.eventItemList.clear()
-        this.eventItemList.addAll(eventItemList)
+    fun setEventItem(myEventItemList: MutableList<Event>) {
+        this.myEventItemList.clear()
+        this.myEventItemList.addAll(myEventItemList)
         notifyDataSetChanged()
     }
 
-    fun setOnEventClickListener(onEventClicked: (event: Event) -> Unit) {
-        this.onEventClicked = onEventClicked
-    }
-
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return ViewHolder(ItemEventBinding.inflate(li), onEventClicked)
+        return ViewHolder(ItemMyEventBinding.inflate(li), onMyEventClicked)
+
     }
 
-    override fun getItemCount(): Int = eventItemList.size
+    override fun getItemCount(): Int = myEventItemList.size
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        return holder.bind(eventItemList[position])
+        return holder.bind(myEventItemList[position])
     }
 
     inner class ViewHolder(
-        private val itemBinding: ItemEventBinding,
-        private val onEventClickListener: (event: Event) -> Unit,
+        private val itemBinding: ItemMyEventBinding,
+        private val onMyEventEventClickListener: (event: Event) -> Unit,
     ) : RecyclerView.ViewHolder(itemBinding.root) {
 
         fun bind(event: Event) {
             itemBinding.cardView.setOnClickListener {
-                onEventClickListener(event)
+                onMyEventEventClickListener(event)
             }
-            itemBinding.tvEventTitle.text = event.eventName
-            itemBinding.tvEventHost.text = event.hostName
+            itemBinding.tvEventName.text = event.eventName
+            itemBinding.tvAddress.text = event.location.addressLine
             itemBinding.tvEventDescription.text = event.description
+            itemBinding.tvEventDateTime.text = event.dateAndTime.toString()
             imageLoader.loadImage(event.photoUrl, itemBinding.imgEvent)
         }
     }
