+++ app/src/main/java/com/gathering/android/event/myevent/view/MyEventFragment.kt
@@ -1,4 +1,4 @@
-package com.gathering.android.event.myevent
+package com.gathering.android.event.myevent.view
 
 import android.os.Bundle
 import android.util.Log
@@ -14,6 +14,8 @@ import com.gathering.android.R
 import com.gathering.android.common.getNavigationResultLiveData
 import com.gathering.android.databinding.FrgMyEventBinding
 import com.gathering.android.event.KEY_ARGUMENT_UPDATE_MY_EVENT_LIST
+import com.gathering.android.event.myevent.viewmodel.MyEventViewModel
+import com.gathering.android.event.myevent.viewmodel.MyEventViewState
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
